feat: add fragments
Instead of using constants.py to hold metadata, we've moved all of the
metadata to "fragments". These are JSON files that are included in the
package via the "package data" mechanism. Each product currently has
three fragments: `bands`, `collection`, and `item-properties`.

This will enable easy addition of new products and their metadata
without having to increase the size of the already-large constants.py
file.